Added template and improved game jam entries
The game jam entries table has been improved. The CSS is written in
BEM-notation and themed with the rest of the template that was also
added.

The route for the game jam entries has been moved to the
`/betadev/gamejam/entries` for a more logical and scalable approach. All
games inherited has also been updated. Also began nesting the routes in
folders, with init files for base routes.
