+++ flask/app/routes/core.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, render_template, send_from_directory
+from flask import Blueprint, render_template
 
 from .utils import read_markdown
 
@@ -21,15 +21,3 @@ def about():
 def for_bedrifter():
     content = read_markdown("for-bedrifter")
     return render_template("blank.html", title="For bedrifter", content=content)
-
-@core_bp.route("/game-jam-entries")
-def game_jam_entries():
-    return render_template("game-jam-entries.html")
-
-@core_bp.route("/game-jam-entries/okt23-maks-uflaks/peak-of-gim/<path:filename>")
-def okt23_maks_uflaks_peak_of_gim(filename):
-    return send_from_directory("./game-jam-entries/okt23-maks-uflaks/peak-of-gim/", filename)
-
-@core_bp.route("/game-jam-entries/okt23-maks-uflaks/crime-download/<path:filename>")
-def okt23_maks_uflaks_crime_download(filename):
-    return send_from_directory("./game-jam-entries/okt23-maks-uflaks/crime-download/", filename)
