+++ flask/app/static/css/global.css
@@ -114,13 +114,13 @@ button:hover {
 
 .container__content h1 {
   font-size: 2.5rem;
-  border-bottom: 2px solid black;
+  border-bottom: 3px solid black;
   padding-bottom: 0.3rem;
 }
 
 .container__content h2 {
   font-size: 2rem;
-  border-bottom: 1px solid black;
+  border-bottom: 2px solid black;
   padding-bottom: 0.3rem;
 }
 
@@ -161,9 +161,42 @@ button:hover {
   text-decoration: none;
 }
 
+.container__content input {
+  padding: 0.5rem 0.75rem;
+  border-radius: 4px;
+  border: 2px solid black;
+  margin: 1rem 0;
+}
+
 .container__content a:hover {
   opacity: 0.8;
   text-decoration: underline;
+}
+
+.container__content table {
+  text-align: left;
+  table-layout: fixed;
+  border-collapse: separate;
+  border-spacing: 0;
+  width: 100%;
+  border: 2px solid black;
+  border-radius: 4px;
+  margin: 0.5rem 0;
+}
+
+.container__content th,
+.container__content td {
+  padding: 0.5rem 0.75rem;
+}
+
+.container__content th {
+  background-color: #0085ff;
+  color: #fff;
+  border-bottom: 2px solid black;
+}
+
+.container__content tr:nth-child(odd) {
+  background-color: #f0f5fa;
 }
 
 /* ------------------ */
@@ -212,6 +245,12 @@ button:hover {
   bottom: 0;
 }
 
+/* Game Jam */
+
+.content__searchbar {
+  margin-bottom: 0.5rem;
+  width: 100%;
+}
