--- flask/app/templates/game-jam-entries.html
@@ -1,184 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Game Jam Entries</title>
-  <link rel="stylesheet" href="{{ url_for('static', filename='css/bulma.css') }}" />
-  <link rel="stylesheet" href="{{ url_for('static', filename='css/custom.css') }}" />
-  <link rel="stylesheet" href="https://www.w3schools.com/w3css/4/w3.css" />
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css"
-    integrity="sha512-iecdLmaskl7CVkqkXNQ/ZH/XLlvWZOJyj7Yy7tcenmpD1ypASozpmT/E0iPtmFIB46ZmdtAc9eNBvH0H/ZpiBw=="
-    crossorigin="anonymous" referrerpolicy="no-referrer" />
-  <style>
-    body {
-      font-family: 'Arial', sans-serif;
-      background-color: #f5f5f5;
-    }
-
-    h1 {
-      color: #333;
-      text-align: center;
-      padding: 20px 0;
-    }
-
-    #searchInput {
-      margin-bottom: 10px;
-      display: block;
-      width: 80%;
-      padding: 10px;
-      margin: 20px auto;
-      border-radius: 5px;
-      border: 1px solid #ddd;
-    }
-
-    #gamesTable {
-      width: 80%;
-      margin: auto;
-      border-collapse: collapse;
-    }
-
-    #gamesTable th,
-    #gamesTable td {
-      padding: 10px;
-      text-align: left;
-      border-bottom: 1px solid #ddd;
-    }
-  </style>
-</head>
-
-<body>
-  <h1>Game Jam Selection</h1>
-
-  <input type="text" id="searchInput" onkeyup="filterTable()" placeholder="Search for games..">
-
-  <table id="gamesTable">
-    <thead>
-      <tr>
-        <th>Name</th>
-        <th>Date</th>
-        <th>Theme</th>
-        <th>Votes</th>
-        <th>Link</th>
-      </tr>
-    </thead>
-    <tbody>
-      <tr>
-        <td>Peak of Gim</td>
-        <td>2023-10-27</td>
-        <td>maks uflaks</td>
-        <td>6</td>
-        <td><a href="/game-jam-entries/okt23-maks-uflaks/peak-of-gim/index.html">Play Game</a></td>
-      </tr>
-      <tr>
-        <td>Crime Download</td>
-        <td>2023-10-27</td>
-        <td>maks uflaks</td>
-        <td>?</td>
-        <td><a href="/game-jam-entries/okt23-maks-uflaks/crime-download/index.html">Play Game</a></td>
-      </tr>
-      <tr>
-        <td>Klaus the Klutz</td>
-        <td>2023-10-27</td>
-        <td>maks uflaks</td>
-        <td>?</td>
-        <td><a href="/game-jam-entries/okt23-maks-uflaks/klaus-the-klutz/index.html">Play Game</a></td>
-      </tr>
-      <tr>
-        <td>Monkey Sort Defense</td>
-        <td>2023-10-27</td>
-        <td>maks uflaks</td>
-        <td>9</td>
-        <td>?</td>
-      </tr>
-      <tr>
-        <td>Max Misadventures</td>
-        <td>2023-10-27</td>
-        <td>maks uflaks</td>
-        <td>5</td>
-        <td>?</td>
-      </tr>
-      <tr>
-        <td>Shamaania</td>
-        <td>2023-10-27</td>
-        <td>maks uflaks</td>
-        <td>?</td>
-        <td>?</td>
-      </tr>
-      <tr>
-        <td>Miss Fortune</td>
-        <td>2023-10-27</td>
-        <td>maks uflaks</td>
-        <td>?</td>
-        <td>?</td>
-      </tr>
-      <tr>
-        <td>Astroholic</td>
-        <td>2023-09-15</td>
-        <td>måne</td>
-        <td>5</td>
-        <td>?</td>
-      </tr>
-      <tr>
-        <td>Neil Bullstrong</td>
-        <td>2023-09-15</td>
-        <td>måne</td>
-        <td>?</td>
-        <td>?</td>
-      </tr>
-      <tr>
-        <td>Moon Clicker</td>
-        <td>2023-09-15</td>
-        <td>måne</td>
-        <td>?</td>
-        <td>?</td>
-      </tr>
-      <tr>
-        <td>Sirkelspill</td>
-        <td>2023-09-15</td>
-        <td>måne</td>
-        <td>?</td>
-        <td>?</td>
-      </tr>
-      <tr>
-        <td>Sondre Rocket Game</td>
-        <td>2023-09-15</td>
-        <td>måne</td>
-        <td>?</td>
-        <td>?</td>
-      </tr>
-      <tr>
-        <td>Simon Rocket Game</td>
-        <td>2023-09-15</td>
-        <td>måne</td>
-        <td>?</td>
-        <td>?</td>
-      </tr>
-    </tbody>
-  </table>
-
-  <script>
-    function filterTable() {
-      var input, filter, table, tr, td, i, txtValue;
-      input = document.getElementById("searchInput");
-      filter = input.value.toUpperCase();
-      table = document.getElementById("gamesTable");
-      tr = table.getElementsByTagName("tr");
-
-      for (i = 0; i < tr.length; i++) {
-        td = tr[i].getElementsByTagName("td")[0];
-        if (td) {
-          txtValue = td.textContent || td.innerText;
-          if (txtValue.toUpperCase().indexOf(filter) > -1) {
-            tr[i].style.display = "";
-          } else {
-            tr[i].style.display = "none";
-          }
-        }
-      }
-    }
-  </script>
-</body>
-
-</html>
